Made transitions persist
The hover transitions are almost definitely slow because of the alpha masks,
which are likely to be especially slow when "in flow" inside a flex container.

Taking the individual boxes out of flow and morphing them with transform: translate3d
will almost definitely be faster.

One other thing: -webkit-mask-image can have a transition applied to it, which makes
the edge alpha look almost like they're morphing. This works poorly with hover CSS,
because if the mouse is moved away before the transition is complete, it will snap
back to the original image.